split account module fixtures
