+++ saleor/account/tests/fixtures/__init__.py
@@ -1,0 +1,1 @@
+from .user import *  # noqa: F403
