+++ saleor/graphql/account/tests/fixtures/__init__.py
@@ -1,0 +1,1 @@
+from .benchmark import *  # noqa: F403
